backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-multipart==0.0.20
pillow==11.1.0
numpy==1.26.4
opencv-python-headless==4.10.0.84
torch==2.4.1
torchvision==0.19.1
transformers==4.46.3
google-generativeai==0.8.3
supabase==2.10.0
python-dotenv==1.0.1
scipy==1.14.1
pytest==8.3.4
httpx==0.27.2
